requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
build==1.2.1
CacheControl==0.14.0
certifi==2024.7.4
charset-normalizer==3.3.2
cleo==2.1.0
colorama==0.4.6
crashtest==0.4.1
decorator==4.4.2
distlib==0.3.8
dulwich==0.21.7
exceptiongroup==1.2.2
fastjsonschema==2.20.0
filelock==3.15.4
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib_metadata==8.0.0
installer==0.7.0
jaraco.classes==3.4.0
keyring==24.3.1
more-itertools==10.3.0
moviepy==1.0.3
msgpack==1.0.8
multidict==6.0.5
numpy==2.0.0
opencv-python==4.10.0.84
packaging==24.1
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.11.1
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
proglog==0.1.10
ptyprocess==0.7.0
pydantic==2.2.1
pydantic_core==2.6.1
pyproject_hooks==1.1.0
pyTelegramBotAPI==4.21.0
python-dotenv==1.0.1
python-telegram-bot==21.3
pytz==2025.1
pywin32-ctypes==0.2.2
rapidfuzz==3.9.4
requests==2.32.3
requests-toolbelt==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
tomli==2.0.1
tomlkit==0.12.5
tqdm==4.66.4
trove-classifiers==2024.7.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.2.2
virtualenv==20.26.3
waifu2x-ncnn-py==2.0.0
win32-setctime==1.1.0
yarl==1.9.4
zipp==3.19.2
